darknetAutomater/requirements.txt
numpy==1.11.0
matplotlib==2.2.0
Pillow==5.1.0
